requirements.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
fastapi==0.87.0
Flask==2.2.2
flatbuffers==22.10.26
gast==0.4.0
gcloud==0.17.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
h5py==3.7.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jws==0.1.3
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauth2client==3.0.0
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pydantic==1.10.2
pyparsing==3.0.9
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
